refactor: ProfileImagePicker 컴포저블 분리
